Rename debit policy package folder.
